+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/DaoOnlyBatchCrudService.java
@@ -38,7 +38,8 @@ public class DaoOnlyBatchCrudService<K extends Key, E extends Entity<K>> impleme
             @NonNull BatchBaseDao<K, E> dao,
             @NonNull KeyFetcher<K> keyFetcher,
             @NonNull ServiceExceptionMapper sem,
-            @NonNull LogLevel exceptionLogLevel) {
+            @NonNull LogLevel exceptionLogLevel
+    ) {
         this.dao = dao;
         this.keyFetcher = keyFetcher;
         this.sem = sem;
@@ -242,15 +243,17 @@ public class DaoOnlyBatchCrudService<K extends Key, E extends Entity<K>> impleme
     }
 
     private List<K> internalBatchInsert(List<E> elements) throws Exception {
-        List<K> collect = elements.stream().filter(e -> Objects.nonNull(e.getKey())).map(E::getKey).collect(Collectors.toList());
+        List<K> collect = elements.stream().filter(
+                e -> Objects.nonNull(e.getKey())).map(E::getKey).collect(Collectors.toList()
+        );
         if (!internalNonExists(collect)) {
             throw new ServiceException(ServiceExceptionCodes.ENTITY_EXISTED);
         }
 
-        for (E element : elements) {
-            if (Objects.isNull(element.getKey())) {
-                element.setKey(keyFetcher.fetchKey());
-            }
+        List<E> nonKeyElements = elements.stream().filter(e -> Objects.isNull(e.getKey())).collect(Collectors.toList());
+        List<K> generatedKeys = keyFetcher.batchFetchKey(nonKeyElements.size());
+        for (int i = 0; i < nonKeyElements.size(); i++) {
+            nonKeyElements.get(i).setKey(generatedKeys.get(i));
         }
 
         return dao.batchInsert(elements);